refactor: lottoService 폴더를 app에서 services로 변경
