+++ src/services/lottoService.js
@@ -48,7 +48,7 @@ const lottoService = {
         winningCriteria: WINNING_CRITERIA, // 당첨 조건으로 필요한 숫자 개수
         reward: REWARD, // 당첨 금액
         count: resultCount[index + 1] || 0, // 당첨된 개수 (없으면 0)
-      }),
+      })
     );
   },
 
@@ -58,7 +58,7 @@ const lottoService = {
   calculateTotalReward(formattedResults) {
     return formattedResults.reduce(
       (total, { count, reward }) => total + count * reward,
-      0,
+      0
     );
   },
 };
